// my-ui/src/generated/roblox-api-types.ts
export type RobloxGeneratedClassName = "BasePlayerGui" | "CoreGui" | "PlayerGui" | "StarterGui" | "GuiBase" | "GuiBase2d" | "GuiObject" | "CanvasGroup" | "Frame" | "GuiButton" | "ImageButton" | "TextButton" | "GuiLabel" | "ImageLabel" | "TextLabel" | "ScrollingFrame" | "TextBox" | "VideoFrame" | "ViewportFrame" | "LayerCollector" | "BillboardGui" | "PluginGui" | "DockWidgetPluginGui" | "QWidgetPluginGui" | "ScreenGui" | "GuiMain" | "SurfaceGuiBase" | "AdGui" | "SurfaceGui" | "UIBase" | "UIComponent" | "UIConstraint" | "UIAspectRatioConstraint" | "UISizeConstraint" | "UITextSizeConstraint" | "UICorner" | "UIGradient" | "UILayout" | "UIGridStyleLayout" | "UIGridLayout" | "UIListLayout" | "UIPageLayout" | "UITableLayout" | "UIPadding" | "UIScale" | "UIStroke" | "StyleBase" | "StyleRule" | "StyleSheet" | "StyleDerive" | "StyleLink" | "StylingService"
// ...
export type RobloxEventHandlers<ClassName extends RobloxGeneratedClassName> = { [EventName in keyof RobloxAuthoringEventMap[ClassName]]?: (...arguments_: RobloxAuthoringEventMap[ClassName][EventName] extends unknown[] ? RobloxAuthoringEventMap[ClassName][EventName] : never) => void }

export type RobloxBasePlayerGuiProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
}

export type RobloxCoreGuiProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  "SelectionImageObject"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
}

export type RobloxPlayerGuiProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  "ScreenOrientation"?: RobloxBindingValue<RobloxEnum_ScreenOrientation>
  "SelectionImageObject"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
}

export type RobloxStarterGuiProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  /** @deprecated This Roblox API member is deprecated. */
  "ResetPlayerGuiOnSpawn"?: RobloxBindingValue<boolean>
  "RtlTextSupport"?: RobloxBindingValue<RobloxEnum_RtlTextSupport>
  "ScreenOrientation"?: RobloxBindingValue<RobloxEnum_ScreenOrientation>
  "ShowDevelopmentGui"?: RobloxBindingValue<boolean>
  "VirtualCursorMode"?: RobloxBindingValue<RobloxEnum_VirtualCursorMode>
}

export type RobloxGuiBaseProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
}

export type RobloxGuiBase2dProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "AutoLocalize"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  "RootLocalizationTable"?: RobloxBindingValue<RobloxInstanceReference<'LocalizationTable'>>
  "SelectionBehaviorDown"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorLeft"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorRight"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorUp"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionGroup"?: RobloxBindingValue<boolean>
}

export type RobloxGuiObjectProperties = {
  "Active"?: RobloxBindingValue<boolean>
  "AnchorPoint"?: RobloxBindingValue<RobloxVector2>
  "Archivable"?: RobloxBindingValue<boolean>
  "AutoLocalize"?: RobloxBindingValue<boolean>
  "AutomaticSize"?: RobloxBindingValue<RobloxEnum_AutomaticSize>
  "BackgroundColor3"?: RobloxBindingValue<RobloxColor3>
  "BackgroundTransparency"?: RobloxBindingValue<number>
  "BorderColor3"?: RobloxBindingValue<RobloxColor3>
  "BorderMode"?: RobloxBindingValue<RobloxEnum_BorderMode>
  "BorderSizePixel"?: RobloxBindingValue<number>
  "ClipsDescendants"?: RobloxBindingValue<boolean>
  /** @deprecated This Roblox API member is deprecated. */
  "Draggable"?: RobloxBindingValue<boolean>
  "LayoutOrder"?: RobloxBindingValue<number>
  "Name"?: RobloxBindingValue<string>
  "NextSelectionDown"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionLeft"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionRight"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionUp"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "Position"?: RobloxBindingValue<RobloxUDim2>
  "RootLocalizationTable"?: RobloxBindingValue<RobloxInstanceReference<'LocalizationTable'>>
  "Rotation"?: RobloxBindingValue<number>
  "Selectable"?: RobloxBindingValue<boolean>
  "SelectionBehaviorDown"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorLeft"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorRight"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorUp"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionGroup"?: RobloxBindingValue<boolean>
  "SelectionImageObject"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "SelectionOrder"?: RobloxBindingValue<number>
  "Size"?: RobloxBindingValue<RobloxUDim2>
  "SizeConstraint"?: RobloxBindingValue<RobloxEnum_SizeConstraint>
  "Visible"?: RobloxBindingValue<boolean>
  "ZIndex"?: RobloxBindingValue<number>
}

export type RobloxCanvasGroupProperties = {
  "Active"?: RobloxBindingValue<boolean>
  "AnchorPoint"?: RobloxBindingValue<RobloxVector2>
  "Archivable"?: RobloxBindingValue<boolean>
  "AutoLocalize"?: RobloxBindingValue<boolean>
  "AutomaticSize"?: RobloxBindingValue<RobloxEnum_AutomaticSize>
  "BackgroundColor3"?: RobloxBindingValue<RobloxColor3>
  "BackgroundTransparency"?: RobloxBindingValue<number>
  "BorderColor3"?: RobloxBindingValue<RobloxColor3>
  "BorderMode"?: RobloxBindingValue<RobloxEnum_BorderMode>
  "BorderSizePixel"?: RobloxBindingValue<number>
  "ClipsDescendants"?: RobloxBindingValue<boolean>
  /** @deprecated This Roblox API member is deprecated. */
  "Draggable"?: RobloxBindingValue<boolean>
  "GroupColor3"?: RobloxBindingValue<RobloxColor3>
  "GroupTransparency"?: RobloxBindingValue<number>
  "LayoutOrder"?: RobloxBindingValue<number>
  "Name"?: RobloxBindingValue<string>
  "NextSelectionDown"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionLeft"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionRight"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "NextSelectionUp"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "Position"?: RobloxBindingValue<RobloxUDim2>
  "RootLocalizationTable"?: RobloxBindingValue<RobloxInstanceReference<'LocalizationTable'>>
  "Rotation"?: RobloxBindingValue<number>
  "Selectable"?: RobloxBindingValue<boolean>
  "SelectionBehaviorDown"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorLeft"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorRight"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorUp"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionGroup"?: RobloxBindingValue<boolean>
  "SelectionImageObject"?: RobloxBindingValue<RobloxInstanceReference<'GuiObject'>>
  "SelectionOrder"?: RobloxBindingValue<number>
  "Size"?: RobloxBindingValue<RobloxUDim2>
  "SizeConstraint"?: RobloxBindingValue<RobloxEnum_SizeConstraint>
  "Visible"?: RobloxBindingValue<boolean>
  "ZIndex"?: RobloxBindingValue<number>
}
// ...
export type RobloxLayerCollectorProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "AutoLocalize"?: RobloxBindingValue<boolean>
  "Enabled"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  "ResetOnSpawn"?: RobloxBindingValue<boolean>
  "RootLocalizationTable"?: RobloxBindingValue<RobloxInstanceReference<'LocalizationTable'>>
  "SelectionBehaviorDown"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorLeft"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorRight"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorUp"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionGroup"?: RobloxBindingValue<boolean>
  "ZIndexBehavior"?: RobloxBindingValue<RobloxEnum_ZIndexBehavior>
}
// ...
export type RobloxPluginGuiProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "AutoLocalize"?: RobloxBindingValue<boolean>
  "Enabled"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
  "ResetOnSpawn"?: RobloxBindingValue<boolean>
  "RootLocalizationTable"?: RobloxBindingValue<RobloxInstanceReference<'LocalizationTable'>>
  "SelectionBehaviorDown"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorLeft"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorRight"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionBehaviorUp"?: RobloxBindingValue<RobloxEnum_SelectionBehavior>
  "SelectionGroup"?: RobloxBindingValue<boolean>
  "Title"?: RobloxBindingValue<string>
  "ZIndexBehavior"?: RobloxBindingValue<RobloxEnum_ZIndexBehavior>
}
// ...
export type RobloxUIBaseProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
}

export type RobloxUIComponentProperties = {
  "Archivable"?: RobloxBindingValue<boolean>
  "Name"?: RobloxBindingValue<string>
}
// ...
export type RobloxAuthoringProperties = {
  "BasePlayerGui": RobloxBasePlayerGuiProperties
  "CoreGui": RobloxCoreGuiProperties
  "PlayerGui": RobloxPlayerGuiProperties
  "StarterGui": RobloxStarterGuiProperties
  "GuiBase": RobloxGuiBaseProperties
  "GuiBase2d": RobloxGuiBase2dProperties
  "GuiObject": RobloxGuiObjectProperties
  "CanvasGroup": RobloxCanvasGroupProperties
  "Frame": RobloxFrameProperties
  "GuiButton": RobloxGuiButtonProperties
  "ImageButton": RobloxImageButtonProperties
  "TextButton": RobloxTextButtonProperties
  "GuiLabel": RobloxGuiLabelProperties
  "ImageLabel": RobloxImageLabelProperties
  "TextLabel": RobloxTextLabelProperties
  "ScrollingFrame": RobloxScrollingFrameProperties
  "TextBox": RobloxTextBoxProperties
  "VideoFrame": RobloxVideoFrameProperties
  "ViewportFrame": RobloxViewportFrameProperties
  "LayerCollector": RobloxLayerCollectorProperties
  "BillboardGui": RobloxBillboardGuiProperties
  "PluginGui": RobloxPluginGuiProperties
  "DockWidgetPluginGui": RobloxDockWidgetPluginGuiProperties
  "QWidgetPluginGui": RobloxQWidgetPluginGuiProperties
  "ScreenGui": RobloxScreenGuiProperties
  "GuiMain": RobloxGuiMainProperties
  "SurfaceGuiBase": RobloxSurfaceGuiBaseProperties
  "AdGui": RobloxAdGuiProperties
  "SurfaceGui": RobloxSurfaceGuiProperties
  "UIBase": RobloxUIBaseProperties
  "UIComponent": RobloxUIComponentProperties
  "UIConstraint": RobloxUIConstraintProperties
  "UIAspectRatioConstraint": RobloxUIAspectRatioConstraintProperties
  "UISizeConstraint": RobloxUISizeConstraintProperties
  "UITextSizeConstraint": RobloxUITextSizeConstraintProperties
  "UICorner": RobloxUICornerProperties
  "UIGradient": RobloxUIGradientProperties
  "UILayout": RobloxUILayoutProperties
  "UIGridStyleLayout": RobloxUIGridStyleLayoutProperties
  "UIGridLayout": RobloxUIGridLayoutProperties
  "UIListLayout": RobloxUIListLayoutProperties
  "UIPageLayout": RobloxUIPageLayoutProperties
  "UITableLayout": RobloxUITableLayoutProperties
  "UIPadding": RobloxUIPaddingProperties
  "UIScale": RobloxUIScaleProperties
  "UIStroke": RobloxUIStrokeProperties
  "StyleBase": RobloxStyleBaseProperties
  "StyleRule": RobloxStyleRuleProperties
  "StyleSheet": RobloxStyleSheetProperties
  "StyleDerive": RobloxStyleDeriveProperties
  "StyleLink": RobloxStyleLinkProperties
  "StylingService": RobloxStylingServiceProperties
}
export type RobloxAuthoringInstance<ClassName extends RobloxGeneratedClassName = RobloxGeneratedClassName> = { ClassName: ClassName; Name?: string; Children?: RobloxAuthoringInstance[]; Events?: RobloxEventHandlers<ClassName> } & RobloxAuthoringProperties[ClassName]
export type RobloxAuthoringTree = RobloxAuthoringInstance | RobloxAuthoringInstance[]
export type RobloxCanonicalInstance<ClassName extends RobloxGeneratedClassName = RobloxGeneratedClassName> = Omit<RobloxAuthoringInstance<ClassName>, 'Events'>
export type RobloxCanonicalTree = RobloxCanonicalInstance | RobloxCanonicalInstance[]